Start profiles renderers.
